bugfix in admin-api test env cleanup: store-api crash fixed
issue was in the order of cleanup: docker db was killed which the
store-api depended on for staying alive, store-api would thereafter be
killed but on shutdown it tries to gracefully disconnect from the db.
this would fail because the db already was killed.

fix: first kill the store-api and then the db