create db procedures outside of migrations/, improves readability
this prevents the need for copy pasting entire functions into a new
migration file whenever the function needs to be updated. that approach
had issues because it: A). decreased readability of code, as it wasnt
necessarily obvious that one has to find the latest function variant,
B). decreased visibility of how exactly a function is being changed in
the related PR, because there is no direct code diff view.

now all db procedures are defined in 1 consistent code location:
store-api-server/db_procedures/